+++ crates/miden-objects/src/errors.rs
@@ -566,6 +566,8 @@ pub enum ProvenTransactionError {
         account_id: AccountId,
         update_size: usize,
     },
+    #[error("proven transaction neither changed the account state, nor consumed any notes")]
+    EmptyTransaction,
 }
 
 // PROPOSED BATCH ERROR