+++ crates/miden-objects/src/transaction/proven_tx.rs
@@ -2,7 +2,7 @@ use alloc::{string::ToString, vec::Vec};
 
 use super::{InputNote, ToInputNoteCommitments};
 use crate::{
-    ACCOUNT_UPDATE_MAX_SIZE, ProvenTransactionError,
+    ACCOUNT_UPDATE_MAX_SIZE, EMPTY_WORD, ProvenTransactionError,
     account::delta::AccountUpdateDetails,
     block::BlockNumber,
     note::NoteHeader,
@@ -18,6 +18,13 @@ use crate::{
 
 /// Result of executing and proving a transaction. Contains all the data required to verify that a
 /// transaction was executed correctly.
+///
+/// A proven transaction must not be empty. A transaction is empty if the account state is unchanged
+/// or the number of input notes is zero. This check prevents proving a transaction once and
+/// submitting it to the network many times. Output notes are not considered because they can be
+/// empty (i.e. contain no assets). Otherwise, a transaction with no account state change, no input
+/// notes and one such empty output note could be resubmitted many times to the network and fill up
+/// block space which is a form of DOS attack.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ProvenTransaction {
     /// A unique identifier for the transaction, see [TransactionId] for additional details.
@@ -113,6 +120,8 @@ impl ProvenTransaction {
     ///
     /// Returns an error if:
     /// - The size of the serialized account update exceeds [`ACCOUNT_UPDATE_MAX_SIZE`].
+    /// - The transaction is empty, which is the case if the account state is unchanged or the
+    ///   number of input notes is zero.
     /// - The transaction was executed against a _new_ on-chain account and its account ID does not
     ///   match the ID in the account update.
     /// - The transaction was executed against a _new_ on-chain account and its commitment does not
@@ -129,6 +138,15 @@ impl ProvenTransaction {
         // If the account is on-chain, then the account update details must be present.
         if self.account_id().is_onchain() {
             self.account_update.validate()?;
+
+            // check that either the account state was changed or at least one note was consumed,
+            // otherwise this transaction is empty
+            if self.account_update.initial_state_commitment()
+                == self.account_update.final_state_commitment()
+                && *self.input_notes.commitment() == EMPTY_WORD
+            {
+                return Err(ProvenTransactionError::EmptyTransaction);
+            }
 
             let is_new_account =
                 self.account_update.initial_state_commitment() == Digest::default();
@@ -328,6 +346,8 @@ impl ProvenTransactionBuilder {
     ///   [`MAX_OUTPUT_NOTES_PER_TX`](crate::constants::MAX_OUTPUT_NOTES_PER_TX).
     /// - The vector of output notes contains duplicates.
     /// - The size of the serialized account update exceeds [`ACCOUNT_UPDATE_MAX_SIZE`].
+    /// - The transaction is empty, which is the case if the account state is unchanged or the
+    ///   number of input notes is zero.
     /// - The transaction was executed against a _new_ on-chain account and its account ID does not
     ///   match the ID in the account update.
     /// - The transaction was executed against a _new_ on-chain account and its commitment does not